Adding warning for singular segments on open boundaries
With the current implementation of open boundaries it can happen that
a segment is part of a special boundary but all of its vertices are
considered as solid walls. As this causes issues when particles leave
through this vertices and the mass cannot be redistributed a warning is
given. This warning can be ignored if the special boundary is used for
moving boundaries or similar.